pyproject.toml
[project]
name = "epmltc"
version = "0.1.0"
description = ""
authors = [
    {name = "Pablo Rubiano"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pyodbc (>=5.3.0,<6.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.4.1,<3.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "joblib (>=1.5.3,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "xlrd (>=2.0.2,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "streamlit (>=1.52.2,<2.0.0)",
    "plotly (>=6.5.1,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
